refactor: LadderGameResult 에서 ladder.startXXX 호출부분 제거하고 immutable value 만 처리
