+++ src/test/java/ladder/domain/LadderLevelTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import ladder.domain.LadderLevel;
 import ladder.domain.LevelItem;
