+++ src/test/java/ladder/domain/LevelTest.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import org.junit.jupiter.api.DisplayName;
